Fix test - clippy
